+++ Assets/Scenes/EndlessMode.unity
@@ -170,7 +170,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 5
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1212204851
 GameObject:
@@ -343,12 +343,12 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1215988023}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 1.12, y: -0.022, z: 6}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalPosition: {x: 1.1199951, y: -0.022003174, z: 6}
   m_LocalScale: {x: 17.2306, y: 11.864744, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 4
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &1215988028
 MonoBehaviour:
@@ -857,5 +857,3 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 13298e8f76039874bb0a81d95b8a5804, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  panelOpt1: {fileID: 0}
-  panelOpt2: {fileID: 0}
