+++ Assets/Scenes/StoryMode.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 599ef6c6fe5bdb1409c2ba2a6f18c0a9
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
