+++ Assets/Scripts/BackgroundScroll.cs
@@ -6,11 +6,17 @@ public class BackgroundScroll : MonoBehaviour
 {
     public float speed = 0f;
     private MeshRenderer mesh;
+    private float ScreenHeight, ScreenWidth;
 
     void Start()
     {
         //get the background's texture
         mesh = GetComponent<MeshRenderer>();
+
+        ScreenHeight = Camera.main.orthographicSize * 2;
+        ScreenWidth = ScreenHeight / Screen.height * Screen.width;
+
+        transform.localScale = new Vector3(ScreenWidth, ScreenHeight, 1);
     }
 
     // Update is called once per frame
@@ -19,5 +25,10 @@ public class BackgroundScroll : MonoBehaviour
         //make the background continually scroll
         //scroll speed depends on speed variable
         mesh.material.mainTextureOffset = new Vector2(Time.time * speed, 0f);
+
+        ScreenHeight = Camera.main.orthographicSize * 2;
+        ScreenWidth = ScreenHeight / Screen.height * Screen.width;
+
+        transform.localScale = new Vector3(ScreenWidth, ScreenHeight, 1);
     }
 }
