--- Assets/Scripts/gameController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2f4182c63030e56419f354ccba04dff7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
